requirements.txt
keyboard>=0.13.5
pystray>=0.19.5
Pillow>=10.0.0
pywin32>=306; sys_platform == 'win32'
psutil>=5.9.0
pycryptodome>=3.20.0
appdirs>=1.4.4
